+++ src/diffusers/__init__.py
@@ -725,6 +725,8 @@ else:
             "LongCatImagePipeline",
             "LTX2ConditionPipeline",
             "LTX2DFRPipeline",
+            "LTX2DFRPipelineOutput",
+            "LTX2DFRTemporalRefinePipeline",
             "LTX2HDRPipeline",
             "LTX2ImageToVideoPipeline",
             "LTX2InContextPipeline",
@@ -1572,6 +1574,8 @@ if TYPE_CHECKING or DIFFUSERS_SLOW_IMPORT:
             LongCatImagePipeline,
             LTX2ConditionPipeline,
             LTX2DFRPipeline,
+            LTX2DFRPipelineOutput,
+            LTX2DFRTemporalRefinePipeline,
             LTX2HDRPipeline,
             LTX2ImageToVideoPipeline,
             LTX2InContextPipeline,
