+++ src/diffusers/pipelines/__init__.py
@@ -341,6 +341,8 @@ else:
         "LTX2Pipeline",
         "LTX2ConditionPipeline",
         "LTX2DFRPipeline",
+        "LTX2DFRPipelineOutput",
+        "LTX2DFRTemporalRefinePipeline",
         "LTX2HDRPipeline",
         "LTX2InContextPipeline",
         "LTX2ImageToVideoPipeline",
@@ -804,6 +806,8 @@ if TYPE_CHECKING or DIFFUSERS_SLOW_IMPORT:
         from .ltx2 import (
             LTX2ConditionPipeline,
             LTX2DFRPipeline,
+            LTX2DFRPipelineOutput,
+            LTX2DFRTemporalRefinePipeline,
             LTX2HDRPipeline,
             LTX2ImageToVideoPipeline,
             LTX2InContextPipeline,
