+++ src/diffusers/pipelines/ltx2/__init__.py
@@ -29,6 +29,8 @@ else:
     _import_structure["pipeline_ltx2"] = ["LTX2Pipeline"]
     _import_structure["pipeline_ltx2_condition"] = ["LTX2ConditionPipeline", "LTX2VideoCondition"]
     _import_structure["pipeline_ltx2_dfr"] = ["LTX2DFRPipeline"]
+    _import_structure["pipeline_ltx2_dfr_temporal_refine"] = ["LTX2DFRTemporalRefinePipeline"]
+    _import_structure["pipeline_output"] = ["LTX2DFRPipelineOutput", "LTX2PipelineOutput", "LTX2VideoDecodeOutput"]
     _import_structure["pipeline_ltx2_diffusion_decode"] = ["LTX2VideoDiffusionDecodePipeline"]
     _import_structure["pipeline_ltx2_hdr_lora"] = ["LTX2HDRPipeline", "LTX2HDRReferenceCondition"]
     _import_structure["pipeline_ltx2_ic_lora"] = ["LTX2InContextPipeline", "LTX2ReferenceCondition"]
@@ -51,11 +53,13 @@ if TYPE_CHECKING or DIFFUSERS_SLOW_IMPORT:
         from .pipeline_ltx2 import LTX2Pipeline
         from .pipeline_ltx2_condition import LTX2ConditionPipeline, LTX2VideoCondition
         from .pipeline_ltx2_dfr import LTX2DFRPipeline
+        from .pipeline_ltx2_dfr_temporal_refine import LTX2DFRTemporalRefinePipeline
         from .pipeline_ltx2_diffusion_decode import LTX2VideoDiffusionDecodePipeline
         from .pipeline_ltx2_hdr_lora import LTX2HDRPipeline, LTX2HDRReferenceCondition
         from .pipeline_ltx2_ic_lora import LTX2InContextPipeline, LTX2ReferenceCondition
         from .pipeline_ltx2_image2video import LTX2ImageToVideoPipeline
         from .pipeline_ltx2_latent_upsample import LTX2LatentUpsamplePipeline
+        from .pipeline_output import LTX2DFRPipelineOutput, LTX2PipelineOutput, LTX2VideoDecodeOutput
         from .vocoder import LTX2Vocoder, LTX2VocoderWithBWE
 
 else:
